Add metrics over alias shadowing
