+++ snuba/query/parsing.py
@@ -1,4 +1,4 @@
-from typing import List
+from typing import Dict, Optional
 
 
 class ParsingContext:
@@ -9,10 +9,13 @@ class ParsingContext:
     """
 
     def __init__(self) -> None:
-        self.__alias_cache: List[str] = []
+        self.__alias_cache: Dict[str, str] = {}
 
-    def add_alias(self, alias: str) -> None:
-        self.__alias_cache.append(alias)
+    def add_alias(self, alias: str, expression: str) -> None:
+        self.__alias_cache[alias] = expression
 
     def is_alias_present(self, alias: str) -> bool:
         return alias in self.__alias_cache
+
+    def get_expression_for_alias(self, alias: str) -> Optional[str]:
+        return self.__alias_cache.get(alias)
